refactor: Move classes into separate file
